# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api [Cometa Aspro Sync]",
        "description": "Rest Api by Laravel for Cometa Aspro Sync",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8002/api",
            "description": "Локальный API сервер v1"
        },
        {
            "url": "http://127.0.0.1:8002/api",
            "description": "Продакшен API сервер v1"
        }
    ],
    "paths": {
        "/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход",
                "description": "Возвращает JWT по email и паролю",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный вход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учётные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход",
                "description": "Выход из системы, токен попадает в blacklist",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Обновить токен",
                "description": "По текущему JWT выдаётся новый токен",
                "operationId": "authRefresh",
                "responses": {
                    "200": {
                        "description": "Новый токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Текущий пользователь",
                "description": "Возвращает данные аутентифицированного пользователя",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResource": {
                "title": "Ошибка",
                "description": "Ответ с информацией об ошибке",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "description": "Статус операции",
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "title": "Код",
                        "description": "Код ошибки",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "msg": {
                        "title": "Сообщение",
                        "description": "Текст сообщения об ошибке",
                        "type": "string",
                        "example": "Внутренняя ошибка сервера"
                    }
                },
                "type": "object"
            },
            "MessageResource": {
                "title": "Простое сообщение",
                "description": "Ответ с простым сообщением",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "description": "Статус операции",
                        "type": "boolean",
                        "example": true
                    },
                    "code": {
                        "title": "Код",
                        "description": "Код ответа",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "msg": {
                        "title": "Сообщение",
                        "description": "Текст сообщения",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
            "TokenResource": {
                "title": "Токен авторизации",
                "description": "JWT и время жизни в секундах",
                "properties": {
                    "token": {
                        "description": "JWT",
                        "type": "string",
                        "example": "eyJ0eXAiOiJKV1QiLCJhbGc..."
                    },
                    "expires_in": {
                        "description": "Время жизни токена в секундах",
                        "type": "integer",
                        "example": 3600
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "Пользователь",
                "description": "Данные пользователя",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "apiKey",
                "description": "API ключ для аутентификации. Передать через заголовок X-API-Key, Authorization Bearer или HTTP Only Cookie. В Swagger UI — заголовок X-API-Key.",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
...
}